Feature/i18n (#174)
* feat: add i18n support

This feature includes the core functions added to /internal, English and Hebrew json files, removal of text strings from the html files, updates to the routes and model updates. Also included are tests. Please note, that this feature is waiting on user registration for final hooks.

* fix(i18n): correct flake8 line length errors

* fix(i18n): attempt to fix build crash at get_translations_dicts() and move foreignkey to last column as requested in CR.

Additional minor doc fixes.

* fix(i18n): simplify check of language file path between live and testing

This change hopefully adds this line to the code coverage.

* fix(i18n): replace glob with path.glob

This change reduces the amount of imports needed and also uses one dir path for both tests and live.

* fix(i18n): revert attempted code improvement

* fix(i18n): added missing language_id to placeholder user

* fix(i18n): added simple test for home

* fix(i18n): code review fixes

* fix(i18n): split the html result dict into two sub-dicts, "variables" and "localized" and the creation of the Jinja2TemplateResponse is now handled in internal/utilities.

* fix(i18n): fixed CR request to not use global state for translations

The translations are now handled with @lru_cache and the function is only re-handled when a new language code parameter is used. This should be the last edit (unless any crash) to this version of the translation feature.