fix: reformat code
First pass at cleaning up the project code.
This fix includes reformatting the code, spacing, and import order.

Note that the HTML files had inconsistent indent regarding jinja2 IF and FOR blocks. This push removes the indent from the places it was in, and a future push will restore a consistent indent to all instances.